refactor(screens): rename screens to views
