--- src/screens/Main/Main.js
@@ -1,8 +1,0 @@
-import React from 'react';
-
-// eslint-disable-next-line
-import styles from './Main.scss';
-
-const Main = ({ children }) => <div className="main">{children}</div>;
-
-export default Main;
