--- src/screens/Main/Main.scss
@@ -1,9 +1,0 @@
-@import 'styles/color.scss';
-@import 'styles/color.scss';
-
-.main {
-  position: relative;
-  padding: 52px;
-  background: $light-rosa;
-  min-height: 100vh;
-}
