--- src/screens/Main/index.js
@@ -1,2 +1,0 @@
-import Main from './Main';
-export default Main;
