+++ src/views/BlockDetail.js
